# backend/swagger.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/exchange/market": {
      "post": {
        "description": "Execute a market order directly to Hyperliquid without executor system. Supports reduce-only and TP/SL.",
        "operationId": "ExchangeController_executeMarketOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectMarketOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Market order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute direct market order",
        "tags": [
          "Exchange"
        ]
      }
    },
    "/exchange/limit": {
      "post": {
        "description": "Execute a limit order directly to Hyperliquid without executor system. Supports reduce-only and TP/SL.",
        "operationId": "ExchangeController_executeLimitOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectLimitOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Limit order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute direct limit order",
        "tags": [
          "Exchange"
        ]
      }
    },
    "/exchange/scale": {
      "post": {
        "description": "Execute multiple limit orders in a price range with configurable size skew. Supports reduce-only and TP/SL.",
        "operationId": "ExchangeController_executeScaleOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectScaleOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scale order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute direct scale order",
        "tags": [
          "Exchange"
        ]
      }
    },
    "/exchange/stop-limit": {
      "post": {
        "description": "Execute a stop limit order with trigger price and limit price. Supports reduce-only and TP/SL.",
        "operationId": "ExchangeController_executeStopLimitOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectStopLimitOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stop limit order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute direct stop limit order",
        "tags": [
          "Exchange"
        ]
      }
    },
    "/exchange/stop-market": {
      "post": {
        "description": "Execute a stop market order with trigger price only. Supports reduce-only and TP/SL.",
        "operationId": "ExchangeController_executeStopMarketOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectStopMarketOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stop market order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute direct stop market order",
        "tags": [
          "Exchange"
        ]
      }
    },
    "/exchange/twap": {
      "post": {
        "description": "Execute a Time-Weighted Average Price order with specified running time and randomization.",
        "operationId": "ExchangeController_executeTwapOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectTwapOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TWAP order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute TWAP order",
        "tags": [
          "Exchange"
        ]
      }
    },
    "/exchange/leverage": {
      "post": {
        "description": "Set the leverage amount and margin type (cross or isolated) for a specific trading asset.",
        "operationId": "ExchangeController_setLeverage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLeverageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leverage set successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetLeverageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set leverage for an asset",
        "tags": [
          "Exchange"
        ]
      }
    },
...
    "/executions/{tradingOrderId}": {
      "delete": {
        "operationId": "ExecutionsController_cancelUnifiedExecution",
        "parameters": [
          {
            "name": "tradingOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Executions"
        ]
      }
    }
  },
  "info": {
    "title": "Brother Terminal Trading API",
    "description": "API for managing trading positions and orders on the Brother Terminal platform for Hyperliquid Engine.",
    "version": "3.0.0",
    "contact": {
      "name": "Brother Terminal",
      "url": "https://brother-terminal.xyz",
      "email": "dev@brother-terminal.xyz"
    }
  },
...
  "servers": [
    {
      "url": "https://api.brother-terminal.xyz",
      "description": "Production (Mainnet)"
    },
    {
      "url": "https://api-testnet.brother-terminal.xyz",
      "description": "Testnet"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OrderExecutionResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the order was successful",
            "example": true
          },
          "externalTxnHash": {
            "type": "string",
            "description": "External transaction hash if successful",
            "example": "0x1234567890abcdef..."
          },
          "error": {
            "type": "string",
            "description": "Error message if unsuccessful",
            "example": "Insufficient balance"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "TakeProfitStopLossDto": {
        "type": "object",
        "properties": {
          "takeProfitPrice": {
            "type": "number",
            "description": "Take profit price",
            "example": 3800
          },
          "stopLossPrice": {
            "type": "number",
            "description": "Stop loss price",
            "example": 3200
          }
        }
      },
      "DirectMarketOrderDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "size": {
            "type": "number",
            "description": "Size in USD",
            "example": 1000
          },
          "reduceOnly": {
            "type": "boolean",
            "description": "Reduce only flag",
            "example": false
          },
          "tpSl": {
            "description": "Take profit and stop loss settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/TakeProfitStopLossDto"
              }
            ]
          },
          "isSpot": {
            "type": "boolean",
            "description": "Whether this is a spot order (internal use)",
            "example": false
          }
        },
        "required": [
          "asset",
          "side",
          "size",
          "reduceOnly"
        ]
      },
      "DirectLimitOrderDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "size": {
            "type": "number",
            "description": "Size in USD",
            "example": 1000
          },
          "limitPrice": {
            "type": "number",
            "description": "Limit price",
            "example": 3500
          },
          "reduceOnly": {
            "type": "boolean",
            "description": "Reduce only flag",
            "example": false
          },
          "tpSl": {
            "description": "Take profit and stop loss settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/TakeProfitStopLossDto"
              }
            ]
          }
        },
        "required": [
          "asset",
          "side",
          "size",
          "limitPrice",
          "reduceOnly"
        ]
      },
      "DirectScaleOrderDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "startUsd": {
            "type": "number",
            "description": "Start USD amount",
            "example": 3400
          },
          "endUsd": {
            "type": "number",
            "description": "End USD amount",
            "example": 3600
          },
          "totalSize": {
            "type": "number",
            "description": "Total size in USD to distribute across all orders",
            "example": 1000
          },
          "totalOrders": {
            "type": "number",
            "description": "Total number of orders",
            "example": 5
          },
          "sizeSkew": {
            "type": "number",
            "description": "Size skew (-1 to 1, where -1 skews to smaller sizes first, 1 skews to larger sizes first)",
            "example": 0
          },
          "reduceOnly": {
            "type": "boolean",
            "description": "Reduce only flag",
            "example": false
          },
          "tpSl": {
            "description": "Take profit and stop loss settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/TakeProfitStopLossDto"
              }
            ]
          }
        },
        "required": [
          "asset",
          "side",
          "startUsd",
          "endUsd",
          "totalSize",
          "totalOrders",
          "sizeSkew",
          "reduceOnly"
        ]
      },
      "DirectStopLimitOrderDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "size": {
            "type": "number",
            "description": "Size in USD",
            "example": 1000
          },
          "stopPrice": {
            "type": "number",
            "description": "Stop price (trigger price)",
            "example": 3450
          },
          "limitPrice": {
            "type": "number",
            "description": "Limit price",
            "example": 3500
          },
          "reduceOnly": {
            "type": "boolean",
            "description": "Reduce only flag",
            "example": false
          },
          "tpSl": {
            "description": "Take profit and stop loss settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/TakeProfitStopLossDto"
              }
            ]
          }
        },
        "required": [
          "asset",
          "side",
          "size",
          "stopPrice",
          "limitPrice",
          "reduceOnly"
        ]
      },
      "DirectStopMarketOrderDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "size": {
            "type": "number",
            "description": "Size in USD",
            "example": 1000
          },
          "stopPrice": {
            "type": "number",
            "description": "Stop price (trigger price)",
            "example": 3450
          },
          "reduceOnly": {
            "type": "boolean",
            "description": "Reduce only flag",
            "example": false
          },
          "tpSl": {
            "description": "Take profit and stop loss settings",
            "allOf": [
              {
                "$ref": "#/components/schemas/TakeProfitStopLossDto"
              }
            ]
          }
        },
        "required": [
          "asset",
          "side",
          "size",
          "stopPrice",
          "reduceOnly"
        ]
      },
      "DirectTwapOrderDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "size": {
            "type": "number",
            "description": "Size in USD",
            "example": 1000
          },
          "twapRunningTime": {
            "type": "number",
            "description": "TWAP running time in minutes",
            "example": 60
          },
          "twapRandomize": {
            "type": "boolean",
            "description": "Whether to randomize TWAP execution",
            "example": true
          },
          "reduceOnly": {
            "type": "boolean",
            "description": "Reduce only flag",
            "example": false
          }
        },
        "required": [
          "asset",
          "side",
          "size",
          "twapRunningTime",
          "twapRandomize",
          "reduceOnly"
        ]
      },
      "SetLeverageDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset symbol",
            "example": "ETH"
          },
          "leverage": {
            "type": "number",
            "description": "Leverage amount (1x to 50x depending on asset)",
            "example": 10
          },
          "isCross": {
            "type": "boolean",
            "description": "Whether to use cross margin (true) or isolated margin (false)",
            "example": true
          }
        },
        "required": [
          "asset",
          "leverage",
          "isCross"
        ]
      },
      "SetLeverageResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether leverage was set successfully",
            "example": true
          },
          "error": {
            "type": "string",
            "description": "Error message if unsuccessful",
            "example": "Invalid leverage for asset"
          }
        },
        "required": [
          "success"
        ]
      }
    }
  }
}